moved xml parsing to parsing.xml module
